Content blocker: second draft.
This version includes a built-in rule editor.
See the man page for usage.

This has not undergone any testing; it is not ready
for public release, but some people might want to
play with it. Cheers, my friends.

This also includes a Vi-like leader-key shortcut
system called "keytree". I hope this might be
useful elsewhere. While the content blocker relies
on keytree to work, keytree is entirely independent
of the content blocker.