Add in-browser incremental find.
This works toward an "omnibar" named boredboard.

The new default for Ctrl-F is to open the find interface.
Ctrl-F then toggles whether the search is visible in the window;
Enter exits the inteface, leaving the search highlighting;
Escape clears the search and exits the interface.

Future expansions will add access to visited, marks, and local
files, plus arbitrary go-to-URI, within the same interface.

This commit also adds a new keymap facility: replacekeytree().
This is heavier than setkeytree() et al, but it guarantees that
the given keymap will receive the GDK event. Reconciliation
between these two methods is TODO.